fix minor details in UI and remove unnecesary files
