+++ bin/main
@@ -1,6 +1,6 @@
 #!/usr/bin/env ruby
-require_relative 'player1.rb'
-require_relative 'board.rb'
+require_relative '../lib/player1.rb'
+require_relative '../lib/board.rb'
 
 
 
@@ -12,7 +12,6 @@ _______ _          _______             _______
    | |  | |/ __|______| |/ _` |/ __|______| |/ _ \ / _ \
    | |  | | (__       | | (_| | (__       | | (_) |  __/
    |_|  |_|\___|      |_|\__,_|\___|      |_|\___/ \___|
-   \n
    '
