--- lib/player.rb
@@ -1,6 +1,0 @@
-class Player
-    attr_reader = name
-    def initialize(name = nil)
-        @name = name
-    end
-end
