+++ lib/player1.rb
